refactor: extract IssueExtraction presenters + CompletionLog
Bundles 2 self-contained extractions from IssueExtractionController.swift:

**IssueExtractionPresenters.swift** (85 lines):
- enum IssueExtractionStatusPresenter
- @MainActor final class ManualIssueExtractionStatusPresenter
- @MainActor final class IssueExtractionFailurePresenter
- SHA-256: d75cc646b256fde2c090bbd0d5b94588b335f4342a22ed63500e99e9e6c2359d
  (lines 4-88 of origin/main:IssueExtractionController.swift)

**IssueExtractionCompletionLog.swift** (14 lines):
- struct IssueExtractionCompletionLog
- Located after the main controller class

Applied memory rule "watch for annotations on line immediately preceding
deleted blocks": verified @MainActor at line 90 (for IssueExtractionController
class) is preserved before deleting the presenter range 4-89.

Reduces IssueExtractionController.swift from 233 to 132 lines.

Tests: 667 (unchanged).
